move step to separete file
